+++ libEM/processor.h
@@ -2500,12 +2500,12 @@ outer radius specifies width of Gaussian starting at inner_radius rather than to
 			if (inner_radius_square>0) {
 				if (dist>inner_radius_square) {
 					if (exponent==2.0f) (*pixel) *= exp(-pow(sqrt(dist)-inner_radius,2.0f) / outer_radius_square);
-					else (*pixel) *= exp(-pow(sqrt(dist)-inner_radius,exponent) / pow(outer_radius_square,exponent/2.0f));
+					else (*pixel) *= exp(-pow(sqrt(dist)-inner_radius,exponent) / pow((float)outer_radius_square,exponent/2.0f));
 				}
 			}
 			else {
 				if (exponent==2.0f) (*pixel) *= exp(-dist / outer_radius_square);
-				else (*pixel) *= exp(-pow(dist,exponent/2.0f) / pow(outer_radius_square,exponent/2.0f));
+				else (*pixel) *= exp(-pow(dist,exponent/2.0f) / pow((float)outer_radius_square,exponent/2.0f));
 			}
 		}
 	};
